[panel] Add FormElementContainer to abstract from functionalities
