+++ mc_rtc_rviz_panel/src/FormWidget.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "ClientWidget.h"
+#include "FormElementContainer.h"
 
 namespace mc_rtc_rviz
 {
@@ -12,12 +13,6 @@ struct FormElement;
 struct FormElementHeader;
 
 struct FormWidget;
-
-template<typename T, typename... Args>
-T * make_new(FormWidget * self, const std::string & name, Args &&... args)
-{
-  return new T(self, name, std::forward<Args>(args)...);
-}
 
 struct FormWidget : public ClientWidget
 {
@@ -27,58 +22,12 @@ public:
 
   void update();
 
-  template<typename T, typename... Args>
-  void element(const std::string & name, Args &&... args);
-
-  void add_element(FormElement * element);
+  inline FormElementContainer * container() noexcept { return container_; }
 
 private:
-  void add_element_to_layout(FormElement * element);
-
-  void make_form_layout();
-
-  QVBoxLayout * vlayout_;
-  QWidget * form_;
-  QFormLayout * requiredLayout_;
-  QToolBox * optionalToolBox_;
-  QFormLayout * optionalLayout_;
-  bool changed_ = true;
-  size_t idx_;
-  std::vector<FormElementHeader *> elements_header_;
-  std::vector<FormElement *> elements_;
+  FormElementContainer * container_;
 private slots:
   void released();
 };
-
-} // namespace mc_rtc_rviz
-
-#include "FormElement.h"
-
-namespace mc_rtc_rviz
-{
-
-template<typename T, typename... Args>
-void FormWidget::element(const std::string & name, Args &&... args)
-{
-  if(elements_.size() <= idx_) { changed_ = true; }
-  else
-  {
-    auto & el = elements_[idx_];
-    if(el->type() != typeid(T).hash_code() || el->name() != name)
-    {
-      changed_ = true;
-      elements_.resize(idx_);
-    }
-    else
-    {
-      if(!el->locked()) { static_cast<T *>(el)->changed(std::forward<Args>(args)...); }
-      idx_ += 1;
-      return;
-    }
-  }
-  elements_.push_back(new T(form_, name, std::forward<Args>(args)...));
-  elements_.back()->type(typeid(T).hash_code());
-  idx_ += 1;
-}
 
 } // namespace mc_rtc_rviz
