+++ mc_rtc_rviz_panel/src/Panel.cpp
@@ -883,6 +883,7 @@ void Panel::got_visual(const WidgetId & id, const rbd::parsers::Visual & visual,
 void Panel::got_form(const WidgetId & id)
 {
   auto & form = get_widget<FormWidget>(id);
+  activeForm_ = form.container();
   form.update();
 }
 
@@ -892,8 +893,7 @@ void Panel::got_form_checkbox(const WidgetId & formId,
                               bool def,
                               bool def_from_user)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::Checkbox>(name, required, def, def_from_user);
+  activeForm_->element<form::Checkbox>(name, required, def, def_from_user);
 }
 
 void Panel::got_form_integer_input(const WidgetId & formId,
@@ -902,8 +902,7 @@ void Panel::got_form_integer_input(const WidgetId & formId,
                                    int def,
                                    bool def_from_user)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::IntegerInput>(name, required, def, def_from_user);
+  activeForm_->element<form::IntegerInput>(name, required, def, def_from_user);
 }
 
 void Panel::got_form_number_input(const WidgetId & formId,
@@ -912,8 +911,7 @@ void Panel::got_form_number_input(const WidgetId & formId,
                                   double def,
                                   bool def_from_user)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::NumberInput>(name, required, def, def_from_user);
+  activeForm_->element<form::NumberInput>(name, required, def, def_from_user);
 }
 
 void Panel::got_form_string_input(const WidgetId & formId,
@@ -922,8 +920,7 @@ void Panel::got_form_string_input(const WidgetId & formId,
                                   const std::string & def,
                                   bool def_from_user)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::StringInput>(name, required, def, def_from_user);
+  activeForm_->element<form::StringInput>(name, required, def, def_from_user);
 }
 
 void Panel::got_form_array_input(const WidgetId & formId,
@@ -933,8 +930,7 @@ void Panel::got_form_array_input(const WidgetId & formId,
                                  bool fixed_size,
                                  bool def_from_user)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::NumberArrayInput>(name, required, def, fixed_size, def_from_user);
+  activeForm_->element<form::NumberArrayInput>(name, required, def, fixed_size, def_from_user);
 }
 
 void Panel::got_form_combo_input(const WidgetId & formId,
@@ -944,8 +940,7 @@ void Panel::got_form_combo_input(const WidgetId & formId,
                                  bool send_index,
                                  int def)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::ComboInput>(name, required, values, send_index, def);
+  activeForm_->element<form::ComboInput>(name, required, values, send_index, def);
 }
 
 void Panel::got_form_data_combo_input(const WidgetId & formId,
@@ -954,8 +949,7 @@ void Panel::got_form_data_combo_input(const WidgetId & formId,
                                       const std::vector<std::string> & ref,
                                       bool send_index)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::DataComboInput>(name, required, data_, ref, send_index);
+  activeForm_->element<form::DataComboInput>(name, required, data_, ref, send_index);
 }
 
 void Panel::got_form_point3d_input(const WidgetId & formId,
@@ -965,9 +959,8 @@ void Panel::got_form_point3d_input(const WidgetId & formId,
                                    bool default_from_user,
                                    bool interactive)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::Point3DInput>(name, required, formId, default_, default_from_user, interactive,
-                                   impl_->int_server_);
+  activeForm_->element<form::Point3DInput>(name, required, formId, default_, default_from_user, interactive,
+                                           impl_->int_server_);
 }
 
 void Panel::got_form_rotation_input(const WidgetId & formId,
@@ -977,9 +970,8 @@ void Panel::got_form_rotation_input(const WidgetId & formId,
                                     bool default_from_user,
                                     bool interactive)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::RotationInput>(name, required, formId, default_, default_from_user, interactive,
-                                    impl_->int_server_);
+  activeForm_->element<form::RotationInput>(name, required, formId, default_, default_from_user, interactive,
+                                            impl_->int_server_);
 }
 
 void Panel::got_form_transform_input(const WidgetId & formId,
@@ -989,9 +981,8 @@ void Panel::got_form_transform_input(const WidgetId & formId,
                                      bool default_from_user,
                                      bool interactive)
 {
-  auto & form = get_widget<FormWidget>(formId);
-  form.element<form::TransformInput>(name, required, formId, default_, default_from_user, interactive,
-                                     impl_->int_server_);
+  activeForm_->element<form::TransformInput>(name, required, formId, default_, default_from_user, interactive,
+                                             impl_->int_server_);
 }
 
 void Panel::got_start_plot(uint64_t id, const std::string & title)
